chore(git-hooks): automate installation via composer
This commit automates the Git hooks setup in the development
environment. The goal is to ensure CaptainHook is activated for all
developers without the need for additional manual commands.

The installation instruction was added to the 'post-autoload-dump'
script in composer.json. This ensures that hooks are automatically
configured whenever dependencies are installed or the autoload is
regenerated.

Using the '--force' and '--skip-existing' flags makes the installation
idempotent, updating the necessary scripts without disrupting the
workflow if the hooks are already present in the .git directory.